Cleanup: clarify the wording about JPA and ORM
JPA is a specification that certain ORMs adhere to (e.g. Hibernate
or EclipseLink). These ORMs typically also provide their own public
APIs that can be accessed directly. So the relationship between
the JPA and ORM terms is nuanced: they are not quite the same,
but they are not completely unrelated either. This commit tries
to improve the wording in this area.

Also there are some other small fixes in the CRUD mission docs.