SWARM-1466: fail fast when unable to create default deployment (#587)
Motivation
----------
If I create a WAR deployment (Maven `<packaging>war</packaging>`)
but don't include neither `undertow` nor `jaxrs` fraction, Swarm
won't be able to find the respective `DefaultDeploymentFactory`
(because it's brought in by those fractions). What it does instead
is that it creates and deploys an empty JAR. I strongly believe
that this is wrong -- the deployment process should end with an error.

Modifications
-------------
Instead of creating an empty JAR, `DefaultDeploymentCreator` returns
a `null` archive if it can't find a `DefaultDeploymentFactory`.
The `RuntimeDeployer` can actually already handle such situation,
so the only change there is that when the default deployment type
is `war`, it produces a better error message (with a hint that maybe
the `undertow` or `jaxrs` fraction are missing).

Result
------
Failing fast in case the default deployment can't be created,
instead of deploying an empty JAR (which is guaranteed to _not_
do what the user wants).