SWARM-1497 - Avoid removing non-temporal jar files (#633)
Motivation
----------
As part of the fix for SWARM-1327, some non temporal jar files that are being exploded are being removed, which is wrong. Only exploded jar files should be removed.

Modifications
-------------
Fix code in order to not remove original unexploded jar files that are going to be exploded later.

Result
------
User jar files aren't being removed.