ENTSWM-171: Add fractions tags to ribbon-secured-client. (#650)
* ENTSWM-171: Add fractions tags to ribbon-secured-client.

Motivation
----------
ribbon-secured-client is not included in bom-certified. In fact,
it is not considered a fraction and so it's completely ignored by
fraction registry, therefore it's not added to the fraction-list.txt
and ignored by CertifiedBomMojo.

Modifications
-------------
Added "Service Discovery,Netflix OSS,Security" fraction tags. Also set
the stability level to STABLE.

Result
------
ribbon-secured-client will be listed in bom-certified.

* ENTSWM-171: Make ribbon-secured-client a "functional" fraction.